Chapter 10/Code Snippets/Display Results/requirements.txt
azure-storage-queue
rich
